backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
python-dotenv
chromadb
pytest
pytest-asyncio
beautifulsoup4
lxml
readability-lxml
lxml_html_clean
PyYAML
